More warnings
5 -Wpointer-bool-conversion — decomp match tricks converted to (void) casts
1 -Wfortify-source — bzero with hardcoded size replaced with sizeof
4 -Wformat/-Wformat-security — fixed format specifiers (%zu, PRIu64, "%s")
1 -Wtautological-constant-out-of-range-compare — bool used as counter, changed to s32
1 -Wliteral-conversion — explicit (s32) cast for float-to-int truncation
1 -Wmisleading-indentation — fixed indentation in sl.c
1 -Wvisibility — struct ActorMarker → ActorMarker typedef
1 -Wunsequenced — removed double assignment in ternary

+++ src/LAIR/lair_particle.c
@@ -306,6 +306,6 @@ void func_80391EA8(ActorMarker *marker, ActorMarker *a1, f32 a2[3], f32 f1, f32 
     actor->unk100 = a1;
 
     *(f32 *)(unk + 0x04) = f2 > 0 ? 0.5 : 0.0;
-    *(s32 *)(unk + 0x0C) = f2 < 0 ? 2.5 : 1.0;
+    *(s32 *)(unk + 0x0C) = (s32)(f2 < 0 ? 2.5 : 1.0);
 }
